Add separate docs for rails and react kits
