+++ playbook/app/pb_kits/playbook/pb_user/docs/_user_block_content_subtitle_rails.md
@@ -1,0 +1,1 @@
+The kit accepts block content and whatever is passed as block content will display in the subtitle section.
