+++ playbook/app/pb_kits/playbook/pb_user/docs/_user_block_content_subtitle_react.md
@@ -1,0 +1,1 @@
+The `subtitle` prop accepts strings as well as nodes. Whatever node is passed will display in the subtitle section.
