+++ playbook/app/pb_kits/playbook/pb_user/docs/index.js
@@ -4,4 +4,4 @@ export { default as UserTextOnly } from './_user_text_only.jsx'
 export { default as UserSize } from './_user_size.jsx'
 export { default as UserVerticalSize } from './_user_vertical_size.jsx'
 export { default as UserSubtitle } from './_user_subtitle.jsx'
-export { default as UserBlockContentSubtitle } from './_user_block_content_subtitle.jsx'
+export { default as UserBlockContentSubtitleReact } from './_user_block_content_subtitle_react.jsx'
